Add Open Graph and Twitter images to landing and guides layouts and tests
